Rolling user enlistment sync
Adds event logic to sync permissions for a user when they are enrolled
or unenrolled from a course, as well as when their role in a course
changes. When an event is triggered, the course where the change in
enrollment occured is reprovisioned against its Panopto server, causing
the permissions to be updated.