Rolling sync accommodates users with both creator and publisher permissions.
If a Moodle role is mapped to both the Creator and Publisher roles in
Panopto, that user will now become both a creator and a publisher the
Panopto course. This works through rolling sync as well: users that are
added or have their role changed to one mapped to both creator and
publisher will be added to the Panopto course with both roles.